Change integration pie threshold to be a double and smaller range with finer adjustment.

+++ Lumidex/Features/Plot/IntegrationPiePlotViewModel.cs
@@ -23,7 +23,7 @@ public partial class IntegrationPiePlotViewModel : PlotViewModel
     public partial string? TelescopeName { get; set; }
 
     [ObservableProperty]
-    public partial int CutoffThreshold { get; set; } = 2;
+    public partial double CutoffThreshold { get; set; } = 2;
 
     public IntegrationPiePlotViewModel(IDbContextFactory<LumidexDbContext> dbContextFactory)
     {
